refactor: PawcutCameraView 정리
- PawcutCameraBottomSubView
- PawcutCameraTopSubView

나눠서 정리
